+++ lib/karafka/web/tracking/helpers/error_info.rb
@@ -3,7 +3,9 @@
 module Karafka
   module Web
     module Tracking
+      # Namespace for tracking related helpers
       module Helpers
+        # Module containing some helper methods useful for extracting extra errors info
         module ErrorInfo
           # Extracts the basic error info
           #