+++ lib/karafka/web/ui/pro/controllers/consumers.rb
@@ -21,10 +21,9 @@ module Karafka
             # Consumers list
             def index
               @current_state = Models::State.current!
-              processes_total = Models::Processes.active(@current_state)
-              @counters = Lib::HashProxy.new(@current_state[:stats])
+              @counters = Models::Counters.new(@current_state)
               @processes, @next_page = Lib::PaginateArray.new.call(
-                processes_total,
+                Models::Processes.active(@current_state),
                 @params.current_page
               )